apps/backend/legacy/django/requirements.txt
asgiref==3.8.1
Django==5.1.15
django-cors-headers==4.4.0
djangorestframework==3.15.2
pytz==2021.3
sqlparse==0.5.4
psycopg2==2.9.9
celery==5.2.7
redis==3.5.3
python-dotenv==1.2.2
pyjwt==2.12.0
requests==2.33.0
mozilla-django-oidc==4.0.1
channels==4.1.0
daphne==4.1.2
python-socketio==5.14.0
django_anymail==12.0
django-ratelimit==4.1.0
sentry-sdk[django]>=2.25.0